+++ src/Language/Haskell/TH/Datatype/Internal.hs
@@ -1,3 +1,5 @@
+{-# LANGUAGE CPP #-}
+
 {-|
 Module      : Language.Haskell.TH.Datatype.Internal
 Description : Backwards-compatible interface to reified information about datatypes.
@@ -13,7 +15,11 @@ module Language.Haskell.TH.Datatype.Internal where
 import Language.Haskell.TH.Syntax
 
 eqTypeName :: Name
+#if MIN_VERSION_base(4,9,0)
 eqTypeName = mkNameG_tc "base" "Data.Type.Equality" "~"
+#else
+eqTypeName = mkNameG_tc "ghc-prim" "GHC.Types" "~"
+#endif
 
 -- This is only needed for GHC 7.6-specific bug
 starKindName :: Name