+++ test/Main.hs
@@ -24,7 +24,6 @@ module Main (main) where
 
 #if __GLASGOW_HASKELL__ >= 704
 import Control.Monad (zipWithM_)
-import GHC.Exts (Constraint)
 #endif
 
 import Language.Haskell.TH
@@ -32,87 +31,7 @@ import Language.Haskell.TH.Datatype
 import Language.Haskell.TH.Lib (starK)
 
 import Harness
-
-type Gadt1Int = Gadt1 Int
-
-infixr 6 :**:
-data Gadt1 (a :: *) where
-  Gadtc1 :: Int          -> Gadt1Int
-  Gadtc2 :: (a,a)        -> Gadt1 a
-  (:**:) :: Bool -> Char -> Gadt1 ()     -- This is declared infix
-  (:!!:) :: Char -> Bool -> Gadt1 Double -- This is not
-
-data Adt1 (a :: *) (b :: *) = Adtc1 (a,b) | Bool `Adtc2` Int
-
-data Gadtrec1 a where
-  Gadtrecc1, Gadtrecc2 :: { gadtrec1a :: a, gadtrec1b :: b } -> Gadtrec1 (a,b)
-
-data Equal :: * -> * -> * -> * where
-  Equalc :: (Read a, Show a) => [a] -> Maybe a -> Equal a a a
-
-data Showable :: * where
-  Showable :: Show a => a -> Showable
-
-data R = R1 { field1, field2 :: Int }
-
-data Gadt2 :: * -> * -> * where
-  Gadt2c1 :: Gadt2 a [a]
-  Gadt2c2 :: Gadt2 [a] a
-  Gadt2c3 :: Gadt2 [a] [a]
-
-data VoidStoS (f :: * -> *)
-
-data StrictDemo = StrictDemo Int !Int {-# UNPACK #-} !Int
-
-#if MIN_VERSION_template_haskell(2,7,0)
-
--- Data families
-
-data family DF (a :: *)
-data instance DF (Maybe a) = DFMaybe Int [a]
-
-# if MIN_VERSION_template_haskell(2,8,0)
-data family DF1 (a :: k)
-# else
-data family DF1 (a :: *)
-# endif
-data instance DF1 b = DF1 b
-
-data family Quoted (a :: *)
-
-# if MIN_VERSION_template_haskell(2,8,0)
-data family Poly (a :: k)
-# else
-data family Poly (a :: *)
-# endif
-data instance Poly a = MkPoly
-
-data family GadtFam (a :: *) (b :: *)
-data instance GadtFam c d where
-  MkGadtFam1 :: x   -> y        -> GadtFam y x
-  (:&&:)     :: e   -> f        -> GadtFam [e] f   -- This is declard infix
-  (:^^:)     :: Int -> Int      -> GadtFam Int Int -- This is not
-  MkGadtFam4 :: (Int ~ z) => z  -> GadtFam z z
-  MkGadtFam5 :: (q ~ Char) => q -> GadtFam Bool Bool
-infixl 3 :&&:
-
-data family FamLocalDec1 a
-data family FamLocalDec2 a b c
-#endif
-
-#if __GLASGOW_HASKELL__ >= 704
-type Konst (a :: Constraint) (b :: Constraint) = a
-type PredSyn1 a b = Konst (Show a) (Read b)
-type PredSyn2 a b = Konst (PredSyn1 a b) (Show a)
-type PredSyn3 c   = Int ~ c
-
-data PredSynT =
-    PredSyn1 Int Int => MkPredSynT1 Int
-  | PredSyn2 Int Int => MkPredSynT2 Int
-  | PredSyn3 Int     => MkPredSynT3 Int
-#endif
-
-return [] -- segment type declarations above from refiy below
+import Types
 
 -- | Test entry point. Tests will pass or fail at compile time.
 main :: IO ()
@@ -126,6 +45,7 @@ main =
      recordTest
      voidstosTest
      strictDemoTest
+     recordVanillaTest
 #if MIN_VERSION_template_haskell(2,7,0)
      dataFamilyTest
      ghc78bugTest
@@ -134,6 +54,7 @@ main =
      gadtFamTest
      famLocalDecTest1
      famLocalDecTest2
+     recordFamTest
 #endif
      fixityLookupTest
 #if __GLASGOW_HASKELL__ >= 704
@@ -223,24 +144,13 @@ gadtrec1Test :: IO ()
 gadtrec1Test =
   $(do info <- reifyDatatype ''Gadtrec1
 
-       let a             = VarT (mkName "a")
-           names@[v1,v2] = map mkName ["v1","v2"]
-           [v1K,v2K]     = map (\n -> KindedTV n starK) names
-
-       let con = ConstructorInfo
-                   { constructorName       = 'Gadtrecc1
-                   , constructorVars       = [v1K, v2K]
-                   , constructorContext    =
-                        [equalPred a (AppT (AppT (TupleT 2) (VarT v1)) (VarT v2))]
-                   , constructorFields     = [VarT v1, VarT v2]
-                   , constructorStrictness = [notStrictAnnot, notStrictAnnot]
-                   , constructorVariant    = RecordConstructor ['gadtrec1a, 'gadtrec1b] }
+       let con = gadtRecVanillaCI
 
        validateDI info
          DatatypeInfo
            { datatypeName    = ''Gadtrec1
            , datatypeContext = []
-           , datatypeVars    = [SigT a starK]
+           , datatypeVars    = [SigT (VarT (mkName "a")) starK]
            , datatypeVariant = Datatype
            , datatypeCons    =
                [ con, con { constructorName = 'Gadtrecc2 } ]
@@ -389,6 +299,11 @@ strictDemoTest =
            }
    )
 
+recordVanillaTest :: IO ()
+recordVanillaTest =
+  $(do info <- reifyRecord 'gadtrec1a
+       validateCI info gadtRecVanillaCI)
+
 #if MIN_VERSION_template_haskell(2,7,0)
 dataFamilyTest :: IO ()
 dataFamilyTest =
@@ -511,6 +426,7 @@ gadtFamTest =
                    , constructorFields     = [ConT ''Int, ConT ''Int]
                    , constructorStrictness = [notStrictAnnot, notStrictAnnot]
                    , constructorVariant    = NormalConstructor }
+               , gadtRecFamCI
                , ConstructorInfo
                    { constructorName       = 'MkGadtFam4
                    , constructorVars       = []
@@ -576,6 +492,11 @@ famLocalDecTest2 =
                    , constructorVariant    = RecordConstructor [mkName "fm0", mkName "fm1"] }]
            }
    )
+
+recordFamTest :: IO ()
+recordFamTest =
+  $(do info <- reifyRecord 'famRec1
+       validateCI info gadtRecFamCI)
 #endif
 
 fixityLookupTest :: IO ()
@@ -616,7 +537,7 @@ reifyDatatypeWithConNameTest =
                   , constructorStrictness = []
                   , constructorVariant    = NormalConstructor
                   }
-              , justConstructorInfo
+              , justCI
               ]
           }
    )
@@ -624,4 +545,4 @@ reifyDatatypeWithConNameTest =
 reifyConstructorTest :: IO ()
 reifyConstructorTest =
   $(do info <- reifyConstructor 'Just
-       validateCI info justConstructorInfo)
+       validateCI info justCI)