+++ ChangeLog.md
@@ -5,7 +5,13 @@
 * Add `resolvePredSynonyms`
 * Add `reifyConstructor`, which allows reification of `ConstructorInfo` from
   a constructor name, and `lookupByConstructorName`, which allows directly
-  looking up a `ConstructorInfo` from a `DatatypeInfo` value.
+  looking up a `ConstructorInfo` from a `DatatypeInfo` value for a given
+  constructor `Name`.
+* Augment `reifyDatatype` to be able to look up `DatatypeInfo` from the `Name`
+  of a record selector for one of its constructors. Also add `reifyRecord` for
+  reification of of `ConstructorInfo` from a record name, and
+  `lookupByRecordName`, which allows directly looking up a `ConstructorInfo`
+  from a `DatatypeInfo` value for a given record `Name`.
 * Fix bug that caused `th-abstraction` to fail on GHC 7.0 and 7.2 when passing
   a vanilla constructor name to `reifyDatatype`
 * Make `normalizeDec` and `normalizeCon` more robust with respect to